requirements.txt
asgiref==3.5.2
Django==4.0.4
django-environ==0.8.1
sqlparse==0.4.2
